+++ examples/multiple_configs/test_2.py
@@ -1,0 +1,10 @@
+import pytest
+
+import hardpy
+
+pytestmark = pytest.mark.module_name("HardPy template")
+
+
+@pytest.mark.case_name("Test 2")
+def test_two():
+    assert True