+++ hardpy/hardpy_panel/frontend/src/hardpy_test_view/SuiteList.tsx
@@ -91,6 +91,8 @@ interface Props extends WithTranslation {
   selectionSupported?: boolean;
   measurementDisplay?: boolean;
   manualCollectMode?: boolean; 
+  currentTestConfig?: string;
+  measurementDisplay?: boolean; 
 }
 
 const SECONDS_TO_MILLISECONDS = 1000;