+++ docs/changelog.md
@@ -4,6 +4,13 @@ Versions follow [Semantic Versioning](https://semver.org/): `<major>.<minor>.<pa
 
 * Add manual test selection mode allowing users to choose specific tests to run via new UI checkboxes.
   [[PR-224](https://github.com/everypinio/hardpy/pull/224)]
+* Add multiple test plan configurations.
+  [[PR-225]](https://github.com/everypinio/hardpy/pull/225)
+
+## 0.18.3
+
+* Add support for pytest 9.
+  [[PR-233](https://github.com/everypinio/hardpy/pull/233)]
 
 ## 0.18.2